refactor: add helper to implement subscription abortion in adapters
